Fix payload field not being set when no secret
